+++ src/index.ts
@@ -22,4 +22,4 @@ export type {HaoriRuntime} from './env';
 export default Haori;
 
 // バージョン情報
-export const version = '0.4.1';
+export const version = '0.4.2';